SCAL-1158: Fixing gsissh init with Gsi::SSH.time_limited_io_read
The old code could cause freeze when handling gsissh init timeout exception - it tried to read the output for logging with IO.read which is blocking.
A new method time_limited_io_read was added to allow reading partial content of blocked IO (e.g. to debug in case of init hangs).